[QFEED-141] Fix: SocialLogin 구조 변경
