+++ module-domain/src/main/java/com/wsws/moduledomain/authcontext/social/SocialLoginService.java
@@ -1,0 +1,7 @@
+package com.wsws.moduledomain.authcontext.social;
+
+import com.wsws.moduledomain.authcontext.social.aggregate.SocialLogin;
+
+public interface SocialLoginService {
+    SocialLogin getSocialLoginInfo(String authorizationCode);
+}
