--- module-domain/src/main/java/com/wsws/moduledomain/user/repo/SocialLoginService.java
@@ -1,7 +1,0 @@
-package com.wsws.moduledomain.user.repo;
-
-import com.wsws.moduledomain.user.vo.SocialLoginInfo;
-
-public interface SocialLoginService {
-    SocialLoginInfo getSocialLoginInfo(String authorizationCode);
-}
